fix(backend): keep fake GCS outside testing package
Move filesystem FakeStorageClient to utils.other.fake_gcs so offline
harness selection stays in the runtime image source closure.

Refs: BasedHardware/omi#11703
Failure-Class: none
